Set PiEstimator iterations datatype to long

+++ hama/rootbeer/piestimator/src/at/illecker/hama/rootbeer/examples/piestimator/PiEstimatorCpuThread.java
@@ -15,8 +15,8 @@ public class PiEstimatorCpuThread implements Runnable {
 
 	@Override
 	public void run() {
-		int in = 0;
-		for (int i = 0; i < m_iterations; i++) {
+		long in = 0;
+		for (long i = 0; i < m_iterations; i++) {
 			double x = 2.0 * Math.random() - 1.0, y = 2.0 * Math.random() - 1.0;
 			if ((Math.sqrt(x * x + y * y) < 1.0)) {
 				in++;
